chore: move `GlobalCache` aside
